+++ src/Rn.DnsUpdater.Core/Services/Interfaces/IConfigService.cs
@@ -1,7 +1,6 @@
-using System.Collections.Generic;
 using Rn.DnsUpdater.Core.Config;
 
-namespace Rn.DnsUpdater.Services;
+namespace Rn.DnsUpdater.Core.Services.Interfaces;
 
 public interface IConfigService
 {
