+++ src/Rn.DnsUpdater.Core/Services/Interfaces/IDnsUpdaterService.cs
@@ -1,0 +1,8 @@
+﻿using Rn.DnsUpdater.Core.Config;
+
+namespace Rn.DnsUpdater.Core.Services.Interfaces;
+
+public interface IDnsUpdaterService
+{
+  Task UpdateEntryAsync(DnsUpdaterEntry entry, CancellationToken stoppingToken);
+}
