+++ src/Rn.DnsUpdater.Core/Services/Interfaces/IHeartbeatService.cs
@@ -1,0 +1,6 @@
+﻿namespace Rn.DnsUpdater.Core.Services.Interfaces;
+
+public interface IHeartbeatService
+{
+  Task TickAsync();
+}
