+++ src/Rn.DnsUpdater/Program.cs
@@ -6,7 +6,8 @@ using Microsoft.Extensions.Logging;
 using NLog;
 using NLog.Extensions.Logging;
 using Rn.DnsUpdater.Core.Config;
-using Rn.DnsUpdater.Services;
+using Rn.DnsUpdater.Core.Services;
+using Rn.DnsUpdater.Core.Services.Interfaces;
 using Rn.NetCore.BasicHttp;
 using Rn.NetCore.BasicHttp.Factories;
 using Rn.NetCore.Common.Abstractions;
