--- src/Rn.DnsUpdater/Services/Interfaces/IDnsUpdaterService.cs
@@ -1,10 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using Rn.DnsUpdater.Core.Config;
-
-namespace Rn.DnsUpdater.Services;
-
-public interface IDnsUpdaterService
-{
-  Task UpdateEntryAsync(DnsUpdaterEntry entry, CancellationToken stoppingToken);
-}
