--- src/Rn.DnsUpdater/Services/Interfaces/IHeartbeatService.cs
@@ -1,8 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Rn.DnsUpdater.Services;
-
-public interface IHeartbeatService
-{
-  Task TickAsync();
-}
